% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_cells_percentage.R
\name{active_cells_percentage}
\alias{active_cells_percentage}
\title{Calculate the Percentage of Active Cells Over Time}
\usage{
active_cells_percentage(
  calcium_matrix_binarized,
  binarize = FALSE,
  plot = FALSE
)
}
\arguments{
\item{calcium_matrix_binarized}{A binarized matrix where each row represents a cell and each column represents a timepoint. This matrix can be generated using the `binarize()` function.}

\item{binarize}{A logical value indicating whether to binarize the calcium matrix. If TRUE, the function will apply the `binarize()` function to the calcium_matrix_binarized before calculation. Defaults to FALSE.}

\item{plot}{A logical value indicating whether to generate a plot of the percentage of active cells over time. Defaults to FALSE.}
}
\value{
If `plot` is FALSE, returns a data frame containing the time points and the percentage of active cells. If `plot` is TRUE, returns a ggplot object of the percentage of active cells over time.
}
\description{
This function calculates the percentage of active cells over time in a binarized calcium matrix.
It can optionally plot the percentage of active cells over time.
}
\examples{
calcium_matrix <- matrix(runif(1000), nrow = 10)
result <- active_cells_percentage(calcium_matrix, binarize = TRUE)
plot <- active_cells_percentage(calcium_matrix, binarize = TRUE, plot = TRUE)
}
