% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coactive_cells.R
\name{coactive_cells}
\alias{coactive_cells}
\title{Calculate and Plot Percentage of Coactive Cells Over Time}
\usage{
coactive_cells(binarized_calcium_matrix, binarize = FALSE, plot = FALSE)
}
\arguments{
\item{binarized_calcium_matrix}{A binary matrix where each row represents a cell and each column represents a timepoint.}

\item{binarize}{A logical value indicating whether to binarize the calcium matrix. If TRUE, the matrix will be binarized using binarize(). Defaults to FALSE.}

\item{plot}{A logical value indicating whether to generate a plot of the percentage of coactive cells over time. Defaults to FALSE.}
}
\value{
A data frame showing the percentage of coactive cells at each timepoint.
}
\description{
This function calculates the percentage of coactive cells over time from a binarized calcium matrix.
It returns a data frame with the time points and corresponding percentage of active cells. Optionally, it can also generate a plot.
}
\examples{
calcium_matrix <- matrix(runif(1000), nrow = 10)
coactive_cells.df <- coactive_cells(calcium_matrix, binarize = TRUE, plot = TRUE)
}
