% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees.R
\name{degrees}
\alias{degrees}
\title{Degree Distribution Analysis}
\usage{
degrees(graph, plot = TRUE)
}
\arguments{
\item{graph}{An igraph object representing the network.}

\item{plot}{A logical value indicating whether to generate and return the degree histogram plot. Defaults to TRUE.}
}
\value{
If `plot` is TRUE, a ggplot object representing the degree histogram. If `plot` is FALSE, a numeric value representing the mean degree.
}
\description{
This function calculates the degree of each node in the network, generates a histogram of the degree distribution,
and returns either the plot or the mean degree based on the `plot` argument.
}
\examples{
binarized_calcium_matrix <- matrix(runif(100, 0, 1), nrow = 10, ncol = 10)
graph <- make_network(binarized_calcium_matrix)
result <- degrees(graph, plot = TRUE)
}
