% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_per_min.R
\name{events_per_min}
\alias{events_per_min}
\title{Calculate Event Frequency per Minute}
\usage{
events_per_min(binarized_calcium_matrix, frame_rate, mean_all = FALSE)
}
\arguments{
\item{binarized_calcium_matrix}{A binarized matrix where each row represents a cell and each column represents a timepoint. Can be created with `binarize()`.}

\item{frame_rate}{The frame rate of the calcium imaging data (frames per second).}

\item{mean_all}{A logical value indicating whether to return the mean frequency of events per minute for all cells. Defaults to FALSE.}
}
\value{
A numeric array representing the event frequency per minute for each cell, or a single numeric value representing the mean frequency if `mean_all` is TRUE.
}
\description{
This function calculates the frequency of events per minute for each cell in a binarized calcium matrix.
}
\examples{
binarized_data <- matrix(sample(c(0, 1), 100, replace = TRUE), nrow = 10)
frame_rate <- 30
event_frequency <- events_per_min(binarized_data, frame_rate)
mean_event_frequency <- events_per_min(binarized_data, frame_rate, TRUE)
}
