% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset.R
\name{get_subset}
\alias{get_subset}
\title{Get Subset of Cells Based on Label}
\usage{
get_subset(coordinates)
}
\arguments{
\item{coordinates}{A dataframe containing X, Y, Cell ID and a Label columns.}
}
\value{
A dataframe containing only the rows where Label is 1 or TRUE. If the Label column does not exist or is not properly formatted, an error is raised.
}
\description{
This function checks if the provided coordinates dataframe contains a "Label" column.
The Label column should contain either a combination of 0 and 1 (for FALSE and TRUE) or logical values (FALSE and TRUE).
Cells marked as TRUE or 1 are considered part of the subset.
}
\examples{
# Example coordinates with a Label column
coordinates <- data.frame(X = runif(10), Y = runif(10), Cell = 1:10, Label = c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0))
subset_cells <- get_subset(coordinates)
print(subset_cells)
}
