% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top5pc_variance.R
\name{get_top5pc_variance}
\alias{get_top5pc_variance}
\title{Extract the Cumulative Variance Explained by the Top 5 Principal Components}
\usage{
get_top5pc_variance(pca_result)
}
\arguments{
\item{pca_result}{A list object containing PCA results. It should include the eigenvalues of the PCA, specifically
the `eigenvalues` field with a `cumulative.variance.percent` component.}
}
\value{
A numeric value representing the cumulative variance explained by the first 5 principal components (PCs). If
the `pca_result` is \code{NULL} or does not contain eigenvalues, the function returns \code{NaN}.
}
\description{
This function extracts the cumulative variance explained by the top 5 principal components (PCs) from
a PCA result object. It returns the cumulative variance percentage up to the 5th principal component.
}
\examples{
# Assuming `pca_result` is the output from a PCA analysis
# Example with a simulated pca_result
pca_result <- list(eigenvalues = list(cumulative.variance.percent = c(25, 40, 55, 65, 75, 80)))
get_top5pc_variance(pca_result)

}
