% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_to_nonlabeled_connections.R
\name{labeled_to_nonlabeled_connections}
\alias{labeled_to_nonlabeled_connections}
\title{Calculate Labeled-to-NonLabeled Connections}
\usage{
labeled_to_nonlabeled_connections(network, coordinates)
}
\arguments{
\item{network}{An igraph object representing the network.}

\item{coordinates}{A data frame containing the coordinates and labeling information of the nodes. Must include a "Label" and "Cell" column.}
}
\value{
A numeric value representing the total number of Labeled-to-NonLabeled connections.
}
\description{
This function calculates the total number of edges that connect labeled nodes (any non-zero value in the Label column) to non-labeled nodes (Label == 0).
}
