% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_to_nonlabeled_connections_normalized.R
\name{labeled_to_nonlabeled_connections_normalized}
\alias{labeled_to_nonlabeled_connections_normalized}
\title{Calculate Normalized Labeled-to-NonLabeled Connections}
\usage{
labeled_to_nonlabeled_connections_normalized(network, coordinates)
}
\arguments{
\item{network}{An igraph object representing the network.}

\item{coordinates}{A data frame containing the coordinates and labeling information of the nodes. Must include a "Label" and "Cell" column.}
}
\value{
A numeric value representing the normalized Labeled-to-NonLabeled connections.
}
\description{
This function calculates the number of Labeled-to-NonLabeled connections (where labeled nodes have any non-zero value in the Label column)
and normalizes the result by the mean degree of the network.
}
