% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_network.R
\name{make_network}
\alias{make_network}
\title{Create a Network object from a binarized calcium matrix}
\usage{
make_network(
  binarized_calcium_matrix,
  lag.max = 1,
  correlation_threshold = "none"
)
}
\arguments{
\item{binarized_calcium_matrix}{A binarized matrix where each row represents a cell and each column represents a timepoint. This matrix can be generated using the `binarize()` function.}

\item{lag.max}{The maximum lag to use in the cross-correlation function (CCF). Defaults to 1.}

\item{correlation_threshold}{The threshold value for filtering edges in the network (Pearson's coefficients go from -1 to +1). Set to "none" to disable filtering. Defaults to "none".}
}
\value{
An `igraph` object representing the network of correlated cells.
}
\description{
This function produces a network based on the maximum cross-correlation between cells' calcium activity.
The user can specify the lag for the cross-correlation function and the correlation threshold for filtering edges.
}
\examples{
binarized_calcium_matrix <- matrix(sample(c(0, 1), 1000, replace = TRUE), nrow = 10)
network <- make_network(binarized_calcium_matrix)
network_no_filter <- make_network(binarized_calcium_matrix, correlation_threshold = "none")
}
