% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform PCA on Calcium Imaging Data}
\usage{
pca(calcium_matrix, binarize = TRUE, plot = TRUE)
}
\arguments{
\item{calcium_matrix}{A matrix where each row represents a cell and each column represents a timepoint.}

\item{binarize}{A logical value indicating whether to binarize the calcium matrix before performing PCA. Defaults to TRUE.}

\item{plot}{A logical value indicating whether to plot the scree plot of eigenvalues. Defaults to TRUE.}
}
\value{
If plot = FALSE, returns a list containing the PCA results and eigenvalues. If plot = TRUE, displays the scree plot and does not return anything.
}
\description{
This function performs Principal Component Analysis (PCA) on the calcium imaging data and optionally plots the scree plot of the eigenvalues.
}
\examples{
calcium_matrix <- matrix(runif(1000), nrow = 10)
pca_results <- pca(calcium_matrix, binarize = TRUE, plot = TRUE)
}
