% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSD.plt.R
\name{PSD.plt}
\alias{PSD.plt}
\title{Plot Power Spectral Density (PSD)}
\usage{
PSD.plt(calcium_matrix, binarize = TRUE, frame_rate = 0.5)
}
\arguments{
\item{calcium_matrix}{A matrix where each row represents a cell and each column represents a timepoint.}

\item{binarize}{A logical value indicating whether to binarize the calcium matrix before performing the PSD analysis. Defaults to TRUE.}

\item{frame_rate}{The frame rate of the calcium imaging data in frames per second. Defaults to 0.5 Hz (2 seconds per frame).}
}
\value{
A ggplot object representing the PSD plot.
}
\description{
This function performs a Power Spectral Density (PSD) analysis on a calcium imaging matrix
and returns only the PSD plot. The user can specify whether to binarize the matrix before
the analysis and set the frame rate.
}
\examples{
calcium_matrix <- matrix(runif(1000), nrow = 10)
PSD_plot <- PSD.plt(calcium_matrix, binarize = TRUE, frame_rate = 0.5)
print(PSD_plot)
}
