% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSD.R
\name{PSD}
\alias{PSD}
\title{Power Spectral Density (PSD) Analysis}
\usage{
PSD(calcium_matrix, binarize = TRUE, frame_rate = 0.5, plot = TRUE)
}
\arguments{
\item{calcium_matrix}{A matrix where each row represents a cell and each column represents a timepoint.}

\item{binarize}{A logical value indicating whether to binarize the calcium matrix before performing the PSD analysis. Defaults to TRUE.}

\item{frame_rate}{The frame rate of the calcium imaging data in frames per second. Defaults to 0.5 Hz (2 seconds per frame).}

\item{plot}{A logical value indicating whether to plot the PSD. Defaults to TRUE.}
}
\value{
A data frame containing the PSD values and corresponding frequencies. If `plot` is TRUE, a PSD plot is also displayed.
}
\description{
This function performs a Power Spectral Density (PSD) analysis on a calcium imaging matrix.
The user can specify whether to binarize the matrix before the analysis, set the frame rate,
and choose to either plot the PSD or return the resulting data frame.
}
\examples{
calcium_matrix <- matrix(runif(1000), nrow = 10)
PSD_results <- PSD(calcium_matrix, binarize = TRUE, frame_rate = 0.5, plot = TRUE)
}
