% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Run Analysis Pipeline on Calcium Imaging Data}
\usage{
pipeline(
  calcium_matrix,
  coordinates,
  dendrogram = FALSE,
  correlation_threshold = 0.3,
  frame_rate = 0.5,
  lag.max = 1,
  big_community_min_members = 5,
  samplename = "sample"
)
}
\arguments{
\item{calcium_matrix}{A matrix where each row represents a cell and each column represents a timepoint.}

\item{coordinates}{A data frame containing X and Y coordinates for each cell. Must include columns "X", "Y", and "Cell".}

\item{dendrogram}{A logical value indicating whether to include a dendrogram in the population activity plot. Defaults to FALSE.}

\item{correlation_threshold}{A numeric value specifying the threshold for filtering edges by weight in the network analysis. Set to "none" to disable filtering. Defaults to 0.3.}

\item{frame_rate}{A numeric value specifying the frame rate (in Hz) for the PSD analysis. Defaults to 0.5.}

\item{lag.max}{A numeric value specifying the lag to be used in the network creation step. Defaults to 1.}

\item{big_community_min_members}{An integer specifying the minimum number of members for a community to be considered "big." Defaults to 5.}

\item{samplename}{A character string specifying the name of the sample. Used to name saved plot images.}
}
\value{
A list containing the results of each analysis step, including plots and calculated features.
}
\description{
This function runs a comprehensive analysis pipeline on calcium imaging data, including normalization, binarization,
population activity plotting, network creation and plotting, PCA analysis, power spectral density (PSD) analysis,
degree distribution analysis, and various network metrics calculations (e.g., clustering coefficient, global efficiency).
}
