% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot a Network Graph from Calcium Data}
\usage{
plot_network(
  graph,
  coordinates,
  label = "communities",
  cell_ID = "none",
  reverse_y_scale = FALSE,
  frequency_values = NULL,
  correlation_threshold = 0.3
)
}
\arguments{
\item{graph}{An igraph object representing the network. Can be created with `make_network()`.}

\item{coordinates}{A data frame containing X and Y coordinates for each cell ID. Must include columns "X", "Y", and "Cell".}

\item{label}{A character string indicating what to label the cells with. Options are "communities" or "frequency". Defaults to "communities".}

\item{cell_ID}{A dataframe of cell IDs (should contain X and Y columns). If set to "none", the nodes will be labeled with their numbers. Defaults to "none".}

\item{reverse_y_scale}{A logical value indicating whether to reverse the Y scale in the plot (useful for matching image coordinates). Defaults to FALSE.}

\item{frequency_values}{A numeric vector containing the frequency of events per minute for each cell. Required if `label = "frequency"`.}

\item{correlation_threshold}{A numeric value specifying the threshold for filtering edges by weight. Set to "none" to disable filtering. Defaults to 0.3.}
}
\value{
A ggplot object representing the network graph.
}
\description{
This function plots a network graph where nodes represent cells and edges represent connections between them.
The nodes can be color-coded based on the selected label, either by community membership or by the frequency of events per minute.
}
\examples{
# Simulate a binarized calcium matrix
binarized_calcium_matrix <- matrix(sample(c(0, 1), 100, replace = TRUE), nrow = 10)

# Generate the network graph
graph <- make_network(binarized_calcium_matrix)

# Simulate XY coordinates for the cells
posXY <- data.frame(X = runif(10), Y = runif(10), Cell = 1:10)

# Simulate frequency values for the cells
frequency_values <- runif(10, 0, 5)

# Plot the network graph with frequency as the label
plot <- plot_network(graph, coordinates = posXY, label = "frequency", frequency_values = frequency_values)
print(plot)
}
